Update to Netlib version of 1999-05-10
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/Netlib_branch@26863 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libf2c/libI77/open.c
@@ -184,7 +184,7 @@ integer f_open(olist *a)
 			opnerr(a->oerr,107,"open")
 		}
 	else
-		sprintf(buf, "fort.%ld", a->ounit);
+		sprintf(buf, "fort.%ld", (long)a->ounit);
 	b->uscrtch = 0;
 	b->uend=0;
 	b->uwrt = 0;
@@ -277,7 +277,7 @@ fk_open(int seq, int fmt, ftnint n)
 #endif
 {	char nbuf[10];
 	olist a;
-	(void) sprintf(nbuf,"fort.%ld",n);
+	(void) sprintf(nbuf,"fort.%ld",(long)n);
 	a.oerr=1;
 	a.ounit=n;
 	a.ofnm=nbuf;
